+++ shared/config/types.ts
@@ -6,7 +6,6 @@ export interface Config {
     smart_waiting: boolean;
     storage_inspection: boolean;
     mouse_humanization: boolean;
-    profiles: boolean;
   };
   security: {
     secure_eval: boolean;